[infra] new comment style
